requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
passlib[bcrypt]
python-jose
pydantic_settings
python-dotenv